HW01 completed Option completed

+++ src/main/java/ru/javawebinar/topjava/web/MealServlet.java
@@ -16,12 +16,14 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Objects;
 
 import static org.slf4j.LoggerFactory.getLogger;
 
 public class MealServlet extends HttpServlet {
     private static final Logger log = getLogger(MealServlet.class);
+
+    private static String INSERT_OR_EDIT = "/update.jsp";
+    private static String LIST_USER = "/meals.jsp";
 
     private MealRepository repository;
 
@@ -35,49 +37,44 @@ public class MealServlet extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         log.debug("GET request");
 
+        String forward = "";
         String action = (request.getParameter("action"));
 
         if (action == null) {
             action = "getall";
+        }
+
+        if (action.equalsIgnoreCase("delete")) {
+            Long mealId = Long.parseLong(request.getParameter("id"));
+            repository.delete(mealId);
+
+            forward = LIST_USER;
+            List<MealTo> meals = MealsUtil.getMealToList(repository.getAll());
+            request.setAttribute("meals", meals);
+            request.setAttribute("formatter", TimeUtil.getDateTimeFormatter());
+        } else if (action.equalsIgnoreCase("update")) {
+            long mealId = Long.parseLong(request.getParameter("id"));
+            Meal meal = repository.getById(mealId);
+
+            forward = INSERT_OR_EDIT;
+            request.setAttribute("meal", meal);
+            request.setAttribute("formatter", TimeUtil.getDateTimeFormatter());
+        } else if (action.equalsIgnoreCase("create")) {
+            Meal meal = new Meal(LocalDateTime.now(), null, 0);
+
+            forward = INSERT_OR_EDIT;
+            request.setAttribute("meal", meal);
+            request.setAttribute("formatter", TimeUtil.getDateTimeFormatter());
+        } else if (action.equalsIgnoreCase("getall")) {
+            forward = LIST_USER;
+            List<MealTo> meals = MealsUtil.getMealToList(repository.getAll());
+            request.setAttribute("meals", meals);
+            request.setAttribute("formatter", TimeUtil.getDateTimeFormatter());
         } else {
-            action.toLowerCase();
+            forward = INSERT_OR_EDIT;
         }
 
-        switch (action) {
-            case "delete": {
-                long mealId = Long.parseLong(request.getParameter("id"));
-                repository.delete(mealId);
-
-                response.sendRedirect("meals"); //проверить и попробовать убрать jsp
-                break;
-            }
-            case "create": {
-                Meal meal = new Meal(LocalDateTime.now(), "default", 0);
-
-                request.setAttribute("meal", meal);
-                request.setAttribute("formatter", TimeUtil.getDateTimeFormatter());
-                request.getRequestDispatcher("/update.jsp").forward(request, response);
-                break;
-            }
-            case "update": {
-                long id = Long.parseLong(Objects.requireNonNull(request.getParameter("id")));
-                Meal meal = repository.getById(id);
-
-                request.setAttribute("meal", meal);
-                request.setAttribute("formatter", TimeUtil.getDateTimeFormatter());
-                request.getRequestDispatcher("/update.jsp").forward(request, response);
-                break;
-            }
-            case "getall":
-            default:
-                List<MealTo> meals = MealsUtil.getMealToList(repository.getAll());
-                request.setAttribute("meals", meals);
-                request.setAttribute("formatter", TimeUtil.getDateTimeFormatter());
-                request.getRequestDispatcher("/meals.jsp").forward(request, response);
-                break;
-        }
-//        request.getRequestDispatcher("/meals.jsp").forward(request, response);
-//        response.sendRedirect("meals.jsp");
+        request.getRequestDispatcher(forward).forward(request, response);
     }
 
     @Override
@@ -85,14 +82,19 @@ public class MealServlet extends HttpServlet {
         log.debug("POST request");
         request.setCharacterEncoding("UTF-8");
 
-        long id = Long.parseLong(Objects.requireNonNull(request.getParameter("id")));
+        String id = request.getParameter("id");
         LocalDateTime dateTime = LocalDateTime.parse(request.getParameter("dateTime"));
         String description = request.getParameter("description");
         int calories = Integer.parseInt(request.getParameter("calories"));
 
-        Meal newMeal = new Meal(id, dateTime, description, calories);
+        Meal newMeal = new Meal(dateTime, description, calories);
 
-        repository.save(newMeal);
+        if (id == null || id.isEmpty()) {
+            repository.addMeal(newMeal);
+        } else {
+            newMeal.setId(Long.parseLong(id));
+            repository.updateMeal(newMeal);
+        }
 
         response.sendRedirect("meals");
     }
